pyproject.toml
[project]
name = "hackthon-0"
version = "0.1.0"
description = "Personal AI Employee - Bronze Tier (Autonomous FTE Hackathon 2026)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "mcp>=1.27.0",
    "google-genai>=1.73.1",
    "openai>=2.32.0",
    "playwright>=1.58.0",
]
